refactor: Reorganize server code. Change path alias.
