+++ apps/www/app/routes/action+/set-theme.ts
@@ -1,4 +1,4 @@
 import { createThemeAction } from '@tipprunde/utils/theme';
-import { themeSession } from '#utils/sessions.server';
+import { themeSession } from '#app/utils/.server/sessions';
 
 export const action = createThemeAction?.(themeSession);
