+++ apps/www/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.17",
   "type": "module",
   "imports": {
-    "#*": "./app/*"
+    "#*": "./*"
   },
   "scripts": {
     "build": "remix vite:build",
