+++ apps/www/tsconfig.json
@@ -4,7 +4,7 @@
   "compilerOptions": {
     // DX
     "paths": {
-      "#*": ["./app/*"]
+      "#*": ["./*"]
     }
   }
 }
